chore: sponsor logos on homepage
