--- src/components/pages/home/sponsors/index.js
@@ -1,3 +1,0 @@
-import Sponsors from './sponsors';
-
-export default Sponsors;
