+++ src/pages/archive.jsx
@@ -5,9 +5,9 @@ import Archive from 'components/pages/archive/archive';
 import Gallery from 'components/pages/archive/gallery';
 import Hero from 'components/pages/archive/hero';
 import Podcast from 'components/pages/archive/podcast';
-import Sponsors from 'components/pages/archive/sponsors';
 import Layout from 'components/shared/layout';
 import SEO from 'components/shared/seo';
+import Sponsors from 'components/shared/sponsors';
 import SEO_DATA from 'constants/seo-data';
 
 const ArchivePage = () => (
