+++ src/pages/index.jsx
@@ -4,11 +4,11 @@ import React from 'react';
 import Hero from 'components/pages/home/hero';
 import Partners from 'components/pages/home/partners';
 // import Speakers from 'components/pages/home/speakers';
-import Sponsors from 'components/pages/home/sponsors';
 import Venue from 'components/pages/home/venue';
 // import Workshops from 'components/pages/home/workshops';
 import Layout from 'components/shared/layout';
 import SEO from 'components/shared/seo';
+import Sponsors from 'components/shared/sponsors';
 
 const HomePage = () => (
   <Layout homepage>
